fix(wasm): strengthen pubsub guard, unify interactive-id, recover state_hash, cleanup on unload (P1)
Four correctness fixes:

- Pub/Sub cycle guard: replace debug_assert! with if-guard +
  tracing::error! + early return, active in release builds.
- interactive-id: unify insert-inline-directive field from option<u64>
  to option<interactive-id> (= option<u32>), removing lossy casts.
- state_hash error recovery: reset hash to HASH_SENTINEL on WASM call
  failure, forcing recollection on the next frame.
- Plugin unload cleanup: track unloaded plugin IDs and drain them in
  the Salsa sync path to clear stale cached contributions.

+++ kasane-wasm/src/convert/display.rs
@@ -44,7 +44,7 @@ pub(crate) fn wit_display_directive_to_directive_with_resolver(
             byte_offset: d.byte_offset as usize,
             content: super::wit_atoms_to_atoms(&d.content),
             interaction: match d.interactive_id {
-                Some(id) => InlineInteraction::Action(InteractiveId::new(id as u32, plugin_tag)),
+                Some(id) => InlineInteraction::Action(InteractiveId::new(id, plugin_tag)),
                 None => InlineInteraction::None,
             },
         },
@@ -176,7 +176,7 @@ pub(crate) fn display_directive_to_wit(directive: &DisplayDirective) -> wit::Dis
             byte_offset: *byte_offset as u32,
             content: super::atoms_to_wit(content),
             interactive_id: match interaction {
-                InlineInteraction::Action(id) => Some(id.local as u64),
+                InlineInteraction::Action(id) => Some(id.local),
                 InlineInteraction::None => None,
             },
         }),
